+++ pkg/telemetryquerylanguage/functions/tqlcommon/func_is_match_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package functions
+package tqlcommon
 
 import (
 	"testing"
