+++ pkg/telemetryquerylanguage/functions/tqlcommon/func_replace_match.go
@@ -12,26 +12,31 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package functions // import "github.com/open-telemetry/opentelemetry-collector-contrib/pkg/telemetryquerylanguage/functions"
+package tqlcommon // import "github.com/open-telemetry/opentelemetry-collector-contrib/pkg/telemetryquerylanguage/functions/tqlcommon"
 
 import (
 	"fmt"
-	"regexp"
+
+	"github.com/gobwas/glob"
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/telemetryquerylanguage/tql"
 )
 
-func IsMatch(target tql.Getter, pattern string) (tql.ExprFunc, error) {
-	regexp, err := regexp.Compile(pattern)
+func ReplaceMatch(target tql.GetSetter, pattern string, replacement string) (tql.ExprFunc, error) {
+	glob, err := glob.Compile(pattern)
 	if err != nil {
-		return nil, fmt.Errorf("the pattern supplied to IsMatch is not a valid regexp pattern: %w", err)
+		return nil, fmt.Errorf("the pattern supplied to replace_match is not a valid pattern: %w", err)
 	}
 	return func(ctx tql.TransformContext) interface{} {
-		if val := target.Get(ctx); val != nil {
-			if valStr, ok := val.(string); ok {
-				return regexp.MatchString(valStr)
+		val := target.Get(ctx)
+		if val == nil {
+			return nil
+		}
+		if valStr, ok := val.(string); ok {
+			if glob.Match(valStr) {
+				target.Set(ctx, replacement)
 			}
 		}
-		return false
+		return nil
 	}, nil
 }
